Even simpler logic for deepclone, cloneinstance
By always setting core to val, we eliminate another branch from
the loop.

This is not only easier to reason about, it's also a gift to the
tracer.  Maybe.  I'm doing it because it's easier to reason about,
and if it's easier for me, it should be easier for the JIT.