Make dynamic functions actual functions
In short, I thought up a method to do this, and have a reason as
well: Grammars distinguish functions and tables as metas for
actions, and consider callable tables to be tables, not functions,
so we can directly call a Node that's supposed to do something.

Also, just in general, a function should be a function unless
there's a reason for it to be a callable table.