Use a queue in table.splice
Before, we were just using table.insert, so the entire tail of the
table was copied over for each insertion, potentially quite
expensive for a long table with many insertions at an early point.

Now, we spill elements onto a queue, and then use push and pop to
reduce this to two moves per relocated element.