Adds (true) hookable functions
These are dynamic functions (i.e. the central function can be
replaced with `patch_dynamic`) which offer pre and posthooks for
extending the functionality.

Included is a session which exercises the core features of the
new functions: hookable, prehook, and posthook, as well as the
use of patch_dynamic on a function so created.