Doc pass + fixes multiple returns
I forgot that `fn(unpack(tab), val)` will only provide one
parameter, because Lua will drop all additional return values
unless a function is called as the last parameter of a given
function call.

I go back and forth on whether the team made the right call here,
to be honest. Sometimes it prevents an ugly bug where
`fn1(fn2(a), b)` can sometimes bump `b` if `fn(2)` has variadic
returns, but this *could* be prevented with `fn1((fn2(a)), b)`,
while the convention which was adopted clearly requires shenanigans
to get the effect I was looking for.

In any case, this performs correctly, and I've added a session
fixing that behavior as well.