Allow string tags, cache coro collections
It isn't so much "allow string tags" as it is "allow string tags
without causing problems if we subsequently yield that string as
a first argument".

The caching is a good idea when we're expecting identical behavior
from functions, since a given `yield` will have a reference
equality to its peers.